Update empty CPU trace message
When an empty trace file is displayed in the
Profiler, it originally would display the following
text: "This trace doesn't contain any data. Please
record a new one and interact with the app to generate
data". This message can be misleading, especially for
users who attained this trace using a profileable build,
as it sounds like the trace was not captured successfully.

Thus, it has been rephrased with a small UI polish
(separated lines for each message to make it clearer),
to indicate that the trace was captured successfully,
but there was no data to collect.

Fixes: 266047421
Test: updated
Change-Id: Iced26292f64fa967bdd29d2379ea7bb6a83e28ac

GitOrigin-RevId: 033b5a727713fc395c064783167650b281d44a91