Use Minimum/RunningArtifactCoordinate in method signatures
Where possible.

LibraryCompatibility requires a generic ArtifactCoordinate still, as
it is first used to query with a minimum version, and then to launch
an inspector with a running version.

InspectorJarTarget requires a generic ArtifactCoordinate but its
Resolved subclass should always be provided with a
RunningArtifactCoordinate in order for it to be passed to
apiServices.launchInspector -- which implies that toWild() should
produce such a RunningArtifactCoordinate for use in ASwB.

Bug: n/a
Test: existing tests adjusted
Change-Id: I6f1234d7a7eb589d9156070039d29d43e942617a

GitOrigin-RevId: 348cdc874aef8d6b46dd5534676df72f298bcf8f