Detect cycle in resource references for KT folding
The ResourceFoldingBuilder for Kotlin doesn't detect and handle cycles
in resource references. This leads to an infinite loop that can cause a
UI freeze.

I've added a regression test for this bug for both Kotlin and Java,
although the Java folding logic already handled the case. Note that the
two languages will display different text in the case of a loop, but
that's not really a problem since such a cycle won't result in a usable
resource at runtime anyway.

Fixes: 295349275
Test: ResourceFoldingBuilderTest, AndroidResourceFoldingTestGenerated
Change-Id: I26c25af9f81c06a2fe700afa1ec7bb19162e2b27

GitOrigin-RevId: b90f69c0941102ad1def58ceb64a5e2b30eda0a8