Add SingleArtifactTaskHandler class
Building on/extending the ProfilerTaskHandler, this abstract
class adds and enforces additional functionality catered towards
tasks backed by a single artifact. It augments the behavior of
the ProfilerTaskHandler by introducing and enforcing the use of
an InterimStage to facilitate the capture of the artifact.

Originally there was a protected method in this class called
findTaskArtifact that was going to serve as a shared helper
method for the children of SingleArtifactTaskHandler. However,
to make it easier to test and reuse, it has been moved to the
TaskHandlerUtils utility class, along with other methods to
assist in task handling.

Also, because state (the stage field) is now added/stored in the
SingleArtifactTaskHandler, and task handler instances live through
the duration of the program, we needed to some cleanup mechanism
to prevent the stage from causing a memory leak. Thus, an exit
method has been added for task handlers to implement, allowing
task handlers to customize the cleanup work they want to perform
before a new task handler is entered.

Bug: 277797528
Test: added
Change-Id: I72618e86722d163449e2dcd5b0dc6cd23d69be29

GitOrigin-RevId: 0e65a9c41fed6f13e96459f3d7bc3ddd412d6234