Add MemoryTaskHandler and implementations
As a child of the SingleArtifactTaskHandler, the
MemoryTaskHandler defines the stage to setup in
order to collect the underlying artifact for the
task.

This change also includes the three implementations
of the MemoryTaskHandler, one for each of the memory
profiler features/configurations the current profiler
provides: heap dump, native allocations, and
java/kotlin allocations. Each of these task handlers
define how to start + stop collecting and load their
respective artifact data, as well as how to construct
their respective TaskArgs. Moreover, each handler
defines what artifacts it can support and a display
name via the getTaskName() method override.

Bug: 277798936
Test: added
Change-Id: I994c3e1344bfb28d05a54c630f5cd648646296c9

GitOrigin-RevId: 2ed45d35218d4a689dbffa115b682b42a28ce6d6