Fix logcat clearing and not refreshing in game tools
In game tools, if we were to resize the logcat window, we would lose all
logs and the only way to get the logs back was to close and open the
logcat again.

The problem was tied to the filtering logic. The filtering logic would
fail with a null-pointer exception due to a missing PSI system when
resolving hyperlinks. Game tools doesn't need a PSI system since it
doesn't have jump to source, not does it focus on Java code.

To address this issue in the smallest and least invasive way, game tools
defines its own HyperlinkDetector with detects nothing. This stops the
hyperlink filter code from executing.

Bug: 289519889
Change-Id: Id7f5a2b1af2eb9676705da3ddc19a12eb4c8302e

GitOrigin-RevId: 9337a81551be2561573d2264ca57dfa506b51a37