Fix MockProject leak in tests
The Kotlin Compiler apparently constructs a MockProject with an
arbitrary parent disposable in some cases, including in our (probably
somewhat weird) test projects.  This causes our test suite leak
checker to complain.  For the simple tests of
GradleDependencyCompletionContributor, we do not need the Kotlin
references, so replace the Reference provider service and mask away
extensions which implicitly depend on the production behaviour of that
service.

Bug: 296525746
Test: this is it
Change-Id: I338347618c19a7130032fac604685718067da889

GitOrigin-RevId: 9c74e9a3b96e06ef6dd196eb44dac23229a7dfad