Eliminate Freeze in UI When View Connection Details
The freeze observed in the attached bug is caused mostly by the
formatting of the code.

It turns out that formatting code causes a freeze in IntelliJ in
general. Loading the html from https://www.google.com/search?q=java
for example into a HTML scratch buffer and invoking the ReformatCode
action will demonstrate this.

To address this issue, we don't format the response if we think it's too
large. When this happens, we add a banner to the view with an action
to reformat manually if desired.

In addition to this, we also switch from creating our own
Editor/Document to using `PsiAwareTextEditorProvider.createEditor()`.
This new way also seems to be faster that the old way.

I also changed the way we format code from using `ReformatCodeProcessor`
to using `CodeStyleManager.reformat()` because it has less overhead.

Note that `ReformatCodeProcessor` runs asynchronously but still hangs
the UI so if offers no advantage.

Note that there is still a small freeze left which I believe is simply
by writing large amounts of data into UI objects which has to be done on
the UI thread.

There is also more significant delay the first time the Details panel is
opened. I'm not sure what the cause for that is.

Fixes: 259650988
Test: Added
Change-Id: If8ac8dda91adc13f93e5d59600bcce786eb6691a

GitOrigin-RevId: ceea31402326ddd5c98bc32e84461dfad3b5f680