Commit changes before deleting an item in property editor.
If a cell value is being edited the editor will commit the changes
when the editor is closed. In this case the property item is being
deleted by the delete button on top of the declared attributes
section. Unfortunately the editor is committing its value after the
value is deleted.

Fix this.

Fixes: 298602871
Test: Added unit test
Change-Id: I27775524309be7023ca440fe588538de0f5169fd

GitOrigin-RevId: 90c66039a77f480f47c8b852ed398006a811e53c