[scratch] [IDEA-319263] do not run scratch run configurations with Gradle/Maven
CompileStepBeforeRun should not depend on Scratch functionality. When CompileStepBeforeRun tries to build a task, the ProjectTaskContext instance already contains a reference to the run configuration. The ProjectTaskContext will be propagated into ProjectTaskRunner#canRun as an additional argument to filter out scratch related tasks at the plugin level. ProjectTaskRunner#canRun(task) and ProjectTaskRunner(project, task) are no longer used inside the Platform and considered as obsolete.

GitOrigin-RevId: 70f5b2b6e6b5e50248531a9159ea48a2b0be66d4