Convert UnusedResourcesProcessor to Kotlin (3/3)
Manual fixes to conversion.

There are a number of further changes I want to make now that this is in
Kotlin and can take advantage of language constructs, but I've tried to
limit changes here to things that hew fairly closely to the original.

The next change I'm making will further address some of those items in a
more scoped way, including changing a number of collections to be
immutable.

Bug: n/a
Test: n/a
Change-Id: I49c9e808c512e365f4d1134c7265c09fa427ed7f

GitOrigin-RevId: 5cfe8e2375105f79d68f74b88516c9b6a4e76dc6