Simplify GutterIconCache.kt
This change just makes a few simplifications and adds some threadsafety
to GutterIconCache by storing the timestamp with the icon together in
a ConcurrentHashMap. We also get rid of the use of "NONE", which was just
to get around the fact that we didn't have a way to distinguish between
a null value and no value.

This also replaces the "getInstance" with a simple field.

Bug: 299481510
Test: GutterIconCacheTest.kt
Change-Id: Ib2359e1fd38336addebe6295125a05c1cef99daf

GitOrigin-RevId: 227c453725b9eabc2414082ec2c47861d13d08c1