Create and pass the task handlers and project to StudioProfilers
The mapping of task types to their respective task handlers
is now declared and instantiated in the AndroidProfilerToolWindow
code. These task handlers need to be accessible in the profiler-side
code, and thus they are passed into the StudioProfilers instance
created in the AndroidProfilerToolWindow.

Because StudioProfilers also needs the Project instance to invoke
the tab creation and opening code, and it is going to be in the
same StudioProfilers constructor as the task handlers, the Project
constructor parameter was also added in this CL.

Bug: 277797525
Bug: 277797528
Test: none
Change-Id: I453607c01dd796fd49be74cdaff9969493bc7277

GitOrigin-RevId: b8f448c93a61b62199bc12a6eddd434192896f1f