Add RecordingListModel data model
This class serves as the model for the list of past and
imported recordings. The user can select any of the listed
recordings to launch a task from.

This model will be the corresponding model for the recording
list componenet in the Compose UI code. It will manage state
of the currently selected recording as well as provide helper
methods to assist in crafting the recording list item content,
such as the list of supported tasks per recording.

Some other changes in this CL are made to refactor existing
utility methods to make them accessible in testing the
RecordingListModel. Also, task handlers creation has been
extracted to a factory class and made a param of the
RecordingListModel to make it easier to test, as usually
the task handlers are instantiated in the tool window code,
which is not easily accessible to fake in testing.

Bug: 277798530
Test: added
Change-Id: Iba700c7bf226982246513245512ae0749702df57

GitOrigin-RevId: ca443d9cfa060b38449a2f4ea7e4d94c792537dd