Add call to open task tab on selected session change
On successful selection of a new session or a terminated session
that was previously selected, the Profiler will now open a
respective tab for the task. This is becaues these two cases of
session selections correspond respectively to the two major cases
in which a task should be launched: launching a task to collect
new data and launching a task using data collected prior.

Because opening the task tab (which then begins the task handler
operations) requires an instance of the Project (to use the
intellij message bus), fully testing this behavior is not feasible.
However, we have enough coverage of the components that make up
the `launchProfilerTask` method to compensate.

Bug: 277797528
Bug: 301164896
Test: added
Change-Id: Ieb0556aa491718b89176bbffccc45a59177af2e4

GitOrigin-RevId: 276ac97921413ee20a35576b226194aacf0368d0