Relax NavXmlIndex heuristic
NavXmlIndex has a heuristic added to address b/230115126 that checks for
a starting "<navigation>" tag before doing further processing on a file,
since we can't check that the parent directory is actually a navigation
directory at indexing time. But that heuristic doesn't appropriately
account for comments at the top of the file, leading to b/300221546.

This change simplifies the heuristic, removing everything except a
simple check for "<navigation". This may result in a few false
positives, but those will be almost immediately handled when the JAXB
deserializer notices that the files have the wrong structure, so nothing
extra should get stored in the index.

On the off chance that there are well-formed navigation files existing
elsewhere in the source tree (other than in resource directories), I'm
adding a check at the time we fetch the file content that it's actually
a resource file. This check used to exist prior to the fix for
b/230115126, just at indexing time when it wasn't allowed. Doing it now
when the index is accessed should be just fine.

Fixes: 300221546
Test: NavXmlIndexTest
Change-Id: I34df557cc65fbbaf8b387f6bbc216eb94fdc7063

GitOrigin-RevId: db38f6f9a15ba9b65b900e26d4b5ee0cd646302e