Use withAndroidModel() in AppInsights tests
The use of onDisk() ends up constructing a project where modules are
associated with the DefaultModuleSystem, which is not sufficiently
realistic to test whether the project is being provided information
about the version control system from the build system, and what to do
if so.

Moving to withAndroidModel() associates the project's module with the
GradleModuleSystem.  Use AndroidProjectBuilder helpers to ensure that
the enableVcsInfo flag is set appropriately for the test.

Bug: 300595679
Test: existing
Change-Id: I30f7a7b9e2ff88668b0482024c5d3dfda5310e7f

GitOrigin-RevId: 7d49b4a2ca51d67f6df05c9e79f233094efbd3ef