DetailsPanelHeader optional support of variants
This change allows users of DetailsPanelHeader to enable or disable the
variants dropdown functionality. Vitals will not show the dropdown
whereas Crashlytics will. Prior to this change, variants dropdown is
shown in the Vitals tab too.

Also contains a small refactor. The code that clears the UI is moved in
separate method.

Test: added a new test.
Bug: 294097863
Change-Id: I2e72cfdde4667b143cc6b32ee49abba9500a71bc

GitOrigin-RevId: e75136dfa7ee9b4dd50f2273f12c83160945fc2d