Add support for HostTests and Device tests IdeArtifacts
Refactor the IDE artifacts into a collection of hostTest
and DeviceTest Artifacts.
This is to prepare for work on AGP side to support the
addition of ScreenshotTest Artifact to support Screenshot
tests, and should make it easier in the future if we were
to add support for more tests on Android Studio.

Bug: 296544990
Test: existing
Change-Id: I96423d7f5f8458ad0f56b198681d0f08af4bc10c

GitOrigin-RevId: 869a2764a7a1d1faeb8488fddee69d19f182b472