Filter unused resources based on file/dir
Choose which files to process for unused resources based on what file(s)
or directory(ies) are selected at the moment, rather than always running
on an entire module.

This represents a change from the existing behavior, but is more in line
with user expectations. If, for example, the user right clicks on a file
in the project view and chooses "Remove Unused Resources", they may
reasonably expect that only resources in that file are removed. The
current behavior, where all resources in that file's module are removed,
is unintuitive.

Fixes: 179166707
Test: UnusedResources(Gradle)?Test
Change-Id: I57d21b28753ab0d0efb5938d04d309f1ac754eea

GitOrigin-RevId: 1906782c1d6f95308ce500c0587bde10750f25f9