Refactor logic for state detection read and cache
This change moves the logic for determining whether a
KtNameReferenceExpression represents a State read into a separate
library. This will allow us to reuse code between a highlighting
extension and a line marker extension. We also take this opportunity
to cache the value so that in the case that we use both of these
extensions, we can share the result between them on any given PSI tree
change.

Bug: 225218822
Test: ComposeStateReadAnnotatorTest.kt
Change-Id: I2981f75110f5c4bfc33641e80397b63b11c4dbd8

GitOrigin-RevId: ce1e4f7ddac98e88e6a8890a3bd5260d43596686