Fix PSD opening for nowinandroid
Issue was in nature of now in android app.
All Android plugins are wrapped with custom ones.
That is the reason Studio cannot find Android
modules right away and we need to wait for sync.
After sync, list is empty as we don't show root and
it's node considered to be collapsed.

Bug: NA
Test: NA
Change-Id: Idf7a3e82841c95e6ae35f50001866bc3b0a74b73

GitOrigin-RevId: d799435c1c207061ea44d66c6f3467562105336d