Add compose tests to adt-ui-compose module
Added the dependenices for testing compose specifically and general
testing to the adt-ui-module. This allowed for the addition of a sample
composable with a corresponding test to show how one can test desktop
compose composable content and behavior.

Due to a bug (github.com/JetBrains/compose-multiplatform/issues/3243)
in skiko causing the instantiation of the ComposeContentTestRule to
crash on Windows, this test class must be ignored on windows (by
configuring the BUILD file) in order to pass pre-submit.

Bug: 272822954
Test: added
Change-Id: Ibc37d8b9c69711ef66e362c2ff5a60e212d84639

GitOrigin-RevId: 8d58c60ac03f68cd5511eb9fb2a04e8983f27100