[Vcs Insights] Map line number whenever there's recorded revision info
...for AQI crash annotation.

Before, we map line number and show gutter icon at the updated
place if the flag (gradle property android.enableVcsInfo) is enabled.
Since we will have a DSL to replace the flag, i.e. no longer single indicator,
we can only try mapping if we can. This might introduce confusions for
users as they find gutter icons are sometimes placed at the the right
position, and sometimes they are not.

Bug: 299196577
Test: updated
Change-Id: I33f7c86ae6c81618a609970c84eec89b43ac63f1

GitOrigin-RevId: 21d74c73bfb86a767fd75ac411edf2eca4c292ad