[PSD] Handle missing dimension in resolved models
At some point the flavor dimension for a product flavor model might
have been nullable, or a missing dimension might have been represented
as the empty string.  It is currently represented by the string
"agp-missing-dimension-for-sync-only", which was treated as the actual
dimension by parts of the PSD.  Handle this special string with the
special attention it deserves.

Fixes: 264013609
Test: manual
Change-Id: I6fad80bdc49bf74729bf2f2ed193db316dffae7a

GitOrigin-RevId: 8c780cbb8e457673c750520a48773eb359cedd0f