Avoid UI freeze in MotionLayoutComponentHelper
The motion editor has a lot of elements that wait for the render thread
in the UI thread. This can cause freezes.
This mitigates the problem in the MotionLayoutComponentHelper by setting
a timeout that avoids the UI thread to be frozen if the action does not
respond on time.
This is not ideal but it is better than causing the freeze.

Fixes: 261173907
Test: N/A
Change-Id: Ifc215f92ca48321707a54890901593434268143c

GitOrigin-RevId: 9625e36be6782f4e518f83eabdcca15a5d6a93ed