Start/stop foreground process detection when embedded LI is enabled/disabled
Calling foregroundProcessDetection.start and stop from
LayoutInspectorManager is a workaround used to prevent foreground
process detection from running in the background even when embedded
LI is not enabled on any device. This won't be necessary when we will
be able to create a new instance of LayoutInspector for each tab in
Running Devices, instead of having a single global instance of
LayoutInspector shared by all the tabs, which is tracked by b/304540563.

Test: LayoutInspectorManagerTest
Bug: 303051313
Change-Id: Iaed18871fb1e5249f51b79bd8a7bb9edc28aa6b1

GitOrigin-RevId: 91118bfc08afe8e53f913e5e31b04ee1f0369676