Rearrange details panel UI
In both Crashlytics and Vitals, rearrange UI labels to reflect mockup.

Most notably, the event and user count labels are moved to the title bar
where the soft wrap action used to be, which is now moved to the right
side of the tabbed pane header.

Device, SDK level, timestamp labels are now associated with the sample
event as opposed to the issue. (SDK Level is no longer a range)

Test: added new test and updated existing.
Bug: 294097863
Bug: 294247767
Fixes: 290647605
Change-Id: I86f36604ddf9a1f49f755ff3e22d350ef4e4c3d7

GitOrigin-RevId: 66145c832aedbd2b51a72fc00cb5e4cd15284379