Avoid NonBlockingReadAction logging unnecessary errors
When a compilation fails, it will throw a LiveEditUpdateException. If
the compilation is running within a NonBlockingReadAction, the action
will automatically log it as an error even if it's handled by the caller
of the action.
This change replaces the NonBlockingReadAction with the coroutines
equivalent that does not have that problem. The flow remains the same
but the exceptions are handled without being logged.

Fixes: 304845152
Test: EmbeddedCompileClientImplTest
Change-Id: I2f83ed794ac2d3028d3957f0a9c94fbf4cb3ac70

GitOrigin-RevId: 2181a922996041ed5f85dc4d7fa3de1bbb074fc4