Refactor DetailsPanelHeader, VariantComboBox, StackTraceConsole
This change is mainly a refactoring change to simplify the code in
DetailsPanelHeader and VariantComboBox. The details panels are no longer
responsible for driving state changes in the VariantComboBox, because
the latter now consumes the state flow.

Stack trace console is refactored to remove its depedency on DataManager
and instead adopt a new state flow.

Test: updated existing tests.
Bug: 294097863
Change-Id: I543b3b2b665c9875a656a03e3a43562eb87ce633

GitOrigin-RevId: 0da71124d929a8ae3c4061909dd59a6aaa30da13