Suspend quality management for UI Check mode
UI Check mode only displays a small number of previews at once, so
quality management is not hugely important. In addition, performing a
rendering leads in this case to also performing an analysis and updating
the Problems window, which is detrimental to the user experience.
This suspends the quality management in UI Check mode, and restarts it
when back to normal preview.

Bug: 303086226
Test: test added
Change-Id: I9f43b71fe5e080b336e60cc51829911ec390f8af

GitOrigin-RevId: c2fff2a4904bb6d8a1e379943a6d9bb2aacd1977