Prioritize proceses with names related to preferred process name
The current sorting logic of the device process list has every device
process in lexiographic order, with the exception of the process with
a name completely matching the preferred process name being at the top
of the list. However, there are instances where there are processes
like "c.g.a.a.messaging" and "c.g.a.a.messaging:rcs". If one of these
are the preferred process, it make sense that other process should
also be at the top of the process list with it.

This change adds that any process whose name, leading up to the first
colon, matches the preferred process name leading up to its first colon
is also considered part of the "preferred processes". These preferred
processes' names themseleves will be sorted lexiographically, with the
process whose name fully matching the preferred process name being
brought to the top. The preferred process names will be moved to the
top of the device process list. After that, the rest of the processes
(non preferred) will be added in lexiographic order.

Bug: 277798461
Test: added
Change-Id: I2931c17341459cf507a4a18aed5546cfb73a5dbb

GitOrigin-RevId: 2c507b20b334bde05685f71557e0c5816ab7b933