Add TaskHomeTabModel and TaskEntranceTabModel abstract class
The TaskEntraceTabModel is an abstract class and is to be extended
by tab UI models allowing the user to select and enter a Profiler
task. It will allow for shared logic (e.g. creation of a TaskGridModel)
between the TaskHomeTabModel and forthcoming PastRecordingTabModel,
as well as enforce the implementation of an on click handler for
the Profiler task button each model's corresponding UI will have.

In this change also comes one implementation of the TaskEntranceModel,
the TaskHomeTabModel. The TaskHomeTabModel serves as the data model
for the task home tab. It owns the process list model to manage the
available processes to the user to select from, as well as current
process selection. It also implements the behavior on start Profiler
task button click, reading the process and Profiler task selection
and using such values to launch the Profiler task.

Bug: 277797248
Test: existing
Change-Id: I8384eacf1c73abd7b4156fdfc9f9b0eb3725ef95

GitOrigin-RevId: 8b33fd16ba05c1a0ed531611d19b9e71e6d2ce9f