Add PastRecordingsTabModel
The PastRecordingsTabModel serves as the data model for
the past recordings tab. It owns the recording list model
to manage the available recordings the user can select
from, as well as current recording selection. It also
implements the behavior on open Profiler task button click,
reading the recording and Profiler task selection and using
such values to launch the Profiler task.

Bug: 304594618
Test: existing
Change-Id: Ie9345d6fd1a1aad468165e9c479075e452890f9b

GitOrigin-RevId: 0adab1c0badcff24981f77e08d777bb79697c30a