Insert Modifier for K2 ComposeModifierCompletionContributor
K2 completion does not provide Modifier instance in any case (based on
K1 tests and comment, this is a behavior different from K1). So K2
ComposeModifierCompletionContributor has to insert Modifier whenever it
is needed. Note that K1 counterpart inserts Modifier only when the
completion character is '\t'. Based on the comment, it relies on the
infrastructure for the Modifier insertion when the completion character
is '\n'. In the case of K2, we let ComposeModifierCompletionContributor
always insert the Modifier when it is needed (not rely on infrastructure
for Modifier insertion).

Bug: 284311057
Test: ComposeModifierCompletionContributorTest.modifierAsArgument
Change-Id: I28940a5953c3634a09e3fa8aa00d88c8d562ea1d

GitOrigin-RevId: a2ca17cb4eb2e3f427323bb17c0b595c00bc9af9