Fix 'Rename' action on Kotlin casing warning
The "Rename" action when hovering over a Kotlin warning (such as
"variable should begin with a lower case letter") is currently broken.
It relies upon the DataContext to tell it what the current PSI element
is, but our SplitEditor can't do that.

This fix forwards such a data request to the contained text editor,
which can respond appropriately to the request.

Fixes: 303199878
Test: Manually verified
Change-Id: I6d0f3241a17bf49d85406e44c811adb364108b8a

GitOrigin-RevId: 641ed3850114776452cdb4e2ad0a0947784e7f5c