Add listeners using add and remove methods
By doing this, when a new listener is added we can update it with
the latest state. This way if a lister is added after the events
are sent, the listener is notified before the new events are sent.

This will be useful in a later CL where the UI needs to be added
and removed on an existing instance of Layout Inspector.

Test: InspectorModelTest
Bug: 304590546
Change-Id: I64b4d1ee743fd63ab6e88225fd6e51fba0b9a8c7

GitOrigin-RevId: 2204d5db974b51e3f214b97079188f99f49ff8e6