Fix UpdateTest
It turns out that the update icon sometimes doesn't show on Windows
within 60 seconds. There is nothing obviously wrong in the logs or
screenshots when this happens, and it doesn't reproduce locally (so
it may be a quirk of our test environment).

The fix is to extend the amount of time we wait to invoke a component.

Fixes: 305290519
Test: UpdateTest
Change-Id: I7ce838b32489eb1f201c1be4c8ee541141a2129a

GitOrigin-RevId: 4d5bb1369fb9859ae08cfddb10d2e15dee9b91a9